fix(edit): close portable constraint authority
