+++ tests/integration/test_diff_snapshot_capture.py
@@ -482,44 +482,3 @@ def test_directory_replacement_captures_deletion_and_untracked_ancestor(
     assert sorted(records) == ["pkg", "pkg/nested.py"]
     assert records["pkg/nested.py"]["new_available"] is False
     assert records["pkg"]["old_available"] is False
-
-
-@POSIX_SNAPSHOT_TEST
-def test_staged_snapshot_constraint_config_comes_from_index_plane(
-    tmp_path: Path,
-) -> None:
-    # PR #1254 review 3765536002: staged constraints are index-plane evidence.
-    root = _repo(tmp_path)
-    config = root / "architectural-constraints.yml"
-    config.write_bytes(b"version: 1\nconstraints: []\n")
-    _git(root, "add", config.name)
-    config.write_bytes(b"version: 1\nconstraints: [invalid-worktree]\n")
-    registry = snapshots.DiffSnapshotRegistry()
-
-    created = registry.create(str(root), "staged", [])
-    consumer, error = registry.acquire(str(created["diff_snapshot_id"]), str(root))
-
-    assert error is None
-    assert consumer is not None
-    assert consumer.snapshot.constraint_config_path == config.name
-    assert consumer.snapshot.constraint_config_data == b"version: 1\nconstraints: []\n"
-    assert consumer.snapshot.staged_config_matches_worktree is False
-    consumer.release()
-
-
-@POSIX_SNAPSHOT_TEST
-def test_staged_snapshot_records_source_plane_divergence(tmp_path: Path) -> None:
-    # PR #1254 review 3765536016: live graphs cannot represent dirty staged sources.
-    root = _repo(tmp_path)
-    (root / "old.py").write_text("value = 2\n")
-    _git(root, "add", "old.py")
-    (root / "old.py").write_text("value = 3\n")
-    registry = snapshots.DiffSnapshotRegistry()
-
-    created = registry.create(str(root), "staged", [])
-    consumer, error = registry.acquire(str(created["diff_snapshot_id"]), str(root))
-
-    assert error is None
-    assert consumer is not None
-    assert consumer.snapshot.staged_source_matches_worktree is False
-    consumer.release()
