+++ tests/unit/mcp/tools/test_edit_facade.py
@@ -24,18 +24,11 @@ Covers all §5 required cases from ``.recon/p0-facade-framework-spec.md``:
 from __future__ import annotations
 
 import asyncio
-from pathlib import Path
 from typing import Any
-from unittest.mock import AsyncMock, patch
+from unittest.mock import AsyncMock
 
 import pytest
 
-import tree_sitter_analyzer.diff_snapshot_registry as snapshots
-from tests.unit._diff_snapshot_support import (
-    POSIX_SNAPSHOT_TEST,
-    install_fake_snapshot_materializer,
-    make_repo,
-)
 from tree_sitter_analyzer.mcp.tools.base_tool import BaseMCPTool
 from tree_sitter_analyzer.mcp.tools.facade_tool import FacadeTool
 
@@ -469,69 +462,10 @@ def test_constraints_action_does_not_leak_action_to_inner(tmp_path: Any) -> None
 # ---------------------------------------------------------------------------
 
 
-def test_edit_annotations_not_read_only() -> None:
-    """edit facade spans mutating-intent actions — readOnlyHint must be False."""
-    from tree_sitter_analyzer.mcp.tools.edit_facade import _EDIT_ANNOTATIONS
-
-    assert _EDIT_ANNOTATIONS["readOnlyHint"] is False, (
-        "edit facade cannot claim readOnlyHint=True (mixed read+mutating-intent actions)"
-    )
-
-
-def test_edit_annotations_not_destructive() -> None:
-    """edit facade suggests/analyses; it does not write files."""
-    from tree_sitter_analyzer.mcp.tools.edit_facade import _EDIT_ANNOTATIONS
-
-    assert _EDIT_ANNOTATIONS["destructiveHint"] is False
-
-
-def test_edit_annotations_all_four_hints_present() -> None:
-    """test_every_tool_declares_mcp_annotations requires all 4 hint keys."""
-    from tree_sitter_analyzer.mcp.tools.edit_facade import _EDIT_ANNOTATIONS
-
-    required = {"readOnlyHint", "destructiveHint", "idempotentHint", "openWorldHint"}
-    assert required.issubset(_EDIT_ANNOTATIONS.keys())
-
-
-def test_edit_facade_definition_includes_annotations() -> None:
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    facade = build_edit_facade(project_root=None)
-    defn = facade.get_tool_definition()
-    assert "annotations" in defn
-    annot = defn["annotations"]
-    assert annot["readOnlyHint"] is False
-    assert annot["destructiveHint"] is False
-
-
 # ---------------------------------------------------------------------------
 # 13. Facade description honesty — ast_diff description uses REAL mode params
 #     (Leg D of issue #529 triple-fix)
 # ---------------------------------------------------------------------------
-
-
-def test_ast_diff_facade_description_uses_real_mode_params() -> None:
-    """Leg D: the ast_diff description in the edit facade must reference the
-    REAL mode signatures (old_file/new_file | old_source/new_source |
-    old_ref/new_ref) and must NOT use the nonexistent 'before, after' params.
-    """
-    from tree_sitter_analyzer.mcp.tools.edit_facade import _EDIT_DESCRIPTION
-
-    # Must contain real param names
-    assert "old_ref" in _EDIT_DESCRIPTION, (
-        "ast_diff facade description must mention 'old_ref' (diff_git signature)"
-    )
-    assert "old_file" in _EDIT_DESCRIPTION or "new_file" in _EDIT_DESCRIPTION, (
-        "ast_diff facade description must mention 'old_file'/'new_file' (diff_files signature)"
-    )
-    assert "old_source" in _EDIT_DESCRIPTION or "new_source" in _EDIT_DESCRIPTION, (
-        "ast_diff facade description must mention 'old_source'/'new_source' (diff_strings signature)"
-    )
-
-    # Must NOT use the nonexistent 'before, after' params
-    assert "before, after" not in _EDIT_DESCRIPTION, (
-        "ast_diff facade description must NOT use nonexistent 'before, after' params"
-    )
 
 
 # ---------------------------------------------------------------------------
@@ -539,67 +473,9 @@ def test_ast_diff_facade_description_uses_real_mode_params() -> None:
 # ---------------------------------------------------------------------------
 
 
-def test_edit_facade_schema_includes_action_and_required() -> None:
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    facade = build_edit_facade(project_root=None)
-    schema = facade.get_tool_schema()
-    props = schema["properties"]
-    assert "action" in props
-    assert "action" in schema.get("required", [])
-    # action enum must list all 8 actions.
-    enum_vals = set(props["action"].get("enum", []))
-    expected = {
-        "safe",
-        "guard",
-        "impact",
-        "refactor",
-        "constraints",
-        "pr",
-        "classify",
-        "ast_diff",
-        "release_snapshot",
-    }
-    assert expected == enum_vals
-
-
-def test_edit_facade_schema_lenient_additional_properties() -> None:
-    """The merged facade schema must be lenient (additionalProperties not False)."""
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    facade = build_edit_facade(project_root=None)
-    schema = facade.get_tool_schema()
-    # The schema must be additionalProperties: True (lenient), not False (strict).
-    assert schema.get("additionalProperties") is True
-
-
 # ---------------------------------------------------------------------------
 # Issue #451 — edit action=pr without pr_url must fail loudly via the facade
 # ---------------------------------------------------------------------------
-
-
-def test_edit_pr_action_missing_pr_url_fails_loudly() -> None:
-    """action=pr without pr_url → success:False, ERROR verdict, not 'No changed files'.
-
-    Regression guard for issue #451: an agent that misnames the param (e.g.
-    uses query= instead of pr_url=) would have the extra param stripped by
-    facade projection, leaving only {mode:pr}. The inner must return an error
-    envelope, not silently fall through to an empty local diff review.
-    """
-    facade, inners = _make_fake_facade()
-    # Replace the fake 'pr' inner with a real CodeGraphPRReviewTool
-    from tree_sitter_analyzer.mcp.tools.codegraph_pr_review_tool import (
-        CodeGraphPRReviewTool,
-    )
-
-    real_pr_inner = CodeGraphPRReviewTool(project_root=None)
-    facade.action_map["pr"] = real_pr_inner
-
-    # mode=pr but no pr_url (simulates post-projection args)
-    result = asyncio.run(facade.execute({"action": "pr", "mode": "pr"}))
-    assert result["success"] is False
-    assert result.get("verdict") == "ERROR"
-    assert "pr_url" in result.get("error", "")
 
 
 if __name__ == "__main__":
@@ -610,421 +486,3 @@ if __name__ == "__main__":
 # Issue #641 — edit facade schema must expose modification_type with enum
 # for action=guard discoverability (extra_public_params, NOT required:[])
 # ---------------------------------------------------------------------------
-
-
-def test_edit_facade_schema_has_modification_type_property() -> None:
-    """Schema must declare modification_type so schema-reading agents see it.
-
-    Before fix: modification_type was only reachable via additionalProperties
-    (invisible to schema inspection). After fix: it appears in properties with
-    the authoritative enum — matching the inner ModificationGuardTool schema.
-    """
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    facade = build_edit_facade(project_root=None)
-    schema = facade.get_tool_schema()
-    props = schema["properties"]
-    assert "modification_type" in props, (
-        "modification_type must be declared in the edit facade's public schema "
-        "(not hidden behind additionalProperties)"
-    )
-
-
-def test_edit_facade_modification_type_has_enum() -> None:
-    """modification_type property must carry the full authoritative enum."""
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-    from tree_sitter_analyzer.mcp.tools.modification_guard_tool import (
-        MODIFICATION_TYPES,
-    )
-
-    facade = build_edit_facade(project_root=None)
-    schema = facade.get_tool_schema()
-    prop = schema["properties"]["modification_type"]
-    assert "enum" in prop, "modification_type must declare an enum"
-    assert set(prop["enum"]) == set(MODIFICATION_TYPES), (
-        "facade modification_type enum must match the inner tool's MODIFICATION_TYPES constant"
-    )
-
-
-def test_edit_facade_modification_type_NOT_in_required() -> None:
-    """modification_type must NOT be in facade required[] (runtime-resolved param).
-
-    LOCKED convention: runtime-required params are described in the description
-    text, not in schema required: [] — this prevents the facade validator from
-    rejecting calls before routing (facade required only lists 'action').
-    """
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    facade = build_edit_facade(project_root=None)
-    schema = facade.get_tool_schema()
-    assert "modification_type" not in schema.get("required", []), (
-        "modification_type must NOT appear in facade required[] "
-        "(runtime-resolved param — locked convention, #397 family)"
-    )
-
-
-def test_edit_facade_guard_description_marks_modification_type_required() -> None:
-    """action=guard description must mark modification_type as required (e.g. with *).
-
-    Before fix: the description listed 'Params: symbol, modification_type,
-    file_path' without any required marker — agents had no signal that omitting
-    modification_type triggers an error on the first call.
-    """
-    from tree_sitter_analyzer.mcp.tools.edit_facade import _EDIT_DESCRIPTION
-
-    # The guard line must mark modification_type as required (trailing * or explicit note)
-    guard_lines = [
-        line for line in _EDIT_DESCRIPTION.splitlines() if "action=guard" in line
-    ]
-    assert guard_lines, "edit facade description must have an action=guard line"
-    guard_line = guard_lines[0]
-    assert (
-        "modification_type*" in guard_line
-        or "modification_type (required" in guard_line
-    ), (
-        f"action=guard description line must mark modification_type as required "
-        f"(e.g. 'modification_type*'); got: {guard_line!r}"
-    )
-
-
-def test_action_pr_without_mode_or_pr_url_fails_loudly() -> None:
-    """Codex P1 (#483): facade action=pr with NO explicit mode must not
-    fall back to the inner's diff default and return empty success.
-
-    ``edit({"action": "pr", "query": "<url>"})`` (typoed param) previously
-    reached the inner without mode → diff mode → success "No changed files".
-    The facade pr route now implies mode=pr, so the pr_url guard fires."""
-    import asyncio
-
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    facade = build_edit_facade(".")
-    result = asyncio.run(
-        facade.execute({"action": "pr", "query": "https://github.com/o/r/pull/1"})
-    )
-    assert result["success"] is False
-    assert "pr_url" in result["error"]
-
-
-def test_action_pr_explicit_diff_mode_still_reaches_diff() -> None:
-    """Direct sub-mode selection stays available through the facade."""
-    import asyncio
-
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    facade = build_edit_facade(".")
-    with patch(
-        "tree_sitter_analyzer.mcp.tools.codegraph_pr_review_tool._get_local_diff",
-        return_value="",
-    ) as get_local_diff:
-        result = asyncio.run(facade.execute({"action": "pr", "mode": "diff"}))
-
-    get_local_diff.assert_called_once_with("diff", ".")
-    # diff mode reviews local changes — must not demand pr_url
-    assert result["success"] is True
-    assert result.get("error") is None or "pr_url" not in str(result.get("error"))
-
-
-@pytest.mark.asyncio
-async def test_edit_impact_preserves_legacy_branch_mode(monkeypatch) -> None:
-    from tree_sitter_analyzer.mcp.tools.change_impact_tool import ChangeImpactTool
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    seen: list[dict[str, object]] = []
-
-    async def fake_execute(self, arguments):
-        seen.append(arguments)
-        return {"success": True}
-
-    monkeypatch.setattr(ChangeImpactTool, "execute", fake_execute)
-    await build_edit_facade(None).execute({"action": "impact", "mode": "branch"})
-
-    assert seen == [{"mode": "branch"}]
-
-
-@pytest.mark.asyncio
-async def test_edit_snapshot_consumer_rejects_conflicting_arguments() -> None:
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    with pytest.raises(ValueError, match="DIFF_SNAPSHOT_CONFLICTING_ARGUMENTS"):
-        await build_edit_facade(None).execute(
-            {
-                "action": "ast_diff",
-                "diff_snapshot_id": "ds",
-                "file_path": "x.py",
-                "old_code": "bad",
-            }
-        )
-
-
-@pytest.mark.asyncio
-async def test_edit_snapshot_consumer_accepts_only_frozen_arguments() -> None:
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    result = await build_edit_facade(None).execute(
-        {
-            "action": "ast_diff",
-            "diff_snapshot_id": "missing",
-            "file_path": "x.py",
-            "output_format": "json",
-        }
-    )
-    assert result["error_code"] == "DIFF_SNAPSHOT_EXPIRED"
-
-
-@POSIX_SNAPSHOT_TEST
-def test_edit_impact_snapshot_opt_in_does_not_write(
-    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-) -> None:
-    import asyncio
-
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    root = make_repo(tmp_path)
-    (root / "old.py").write_text("value = 2\n")
-    from tree_sitter_analyzer.diff_snapshot_capture import ChangedFile
-
-    install_fake_snapshot_materializer(
-        monkeypatch,
-        root,
-        records=[ChangedFile("old.py", "M", True, True, False)],
-        inventory_paths=["old.py"],
-    )
-    before = {path.relative_to(root) for path in root.rglob("*")}
-    facade = build_edit_facade(str(root))
-
-    result = asyncio.run(
-        facade.execute(
-            {
-                "action": "impact",
-                "mode": "diff",
-                "capture_diff_snapshot": True,
-                "output_format": "json",
-            }
-        )
-    )
-
-    assert result["success"] is True
-    assert result["changed_files"] == ["old.py"]
-    assert {path.relative_to(root) for path in root.rglob("*")} == before
-    assert (
-        snapshots.close_route_lease(
-            str(result["diff_snapshot_id"]), str(result["route_lease_id"])
-        )
-        is True
-    )
-
-
-@POSIX_SNAPSHOT_TEST
-def test_edit_impact_rejects_clean_tracked_transient_write_restore(
-    tmp_path: Path, monkeypatch
-) -> None:
-    """Strict impact cannot certify analysis that observed a transient clean file."""
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-    from tree_sitter_analyzer.mcp.tools.utils import change_impact_analysis
-
-    # RFC-0022 P0.2 review 2026-07-01: dependency analysis consumed a clean
-    # tracked transient and certified success after the callback restored it.
-    root = make_repo(tmp_path)
-    changed = root / "old.py"
-    dependency = root / "gone.py"
-    changed.write_text("value = 2\n")
-    original = dependency.read_bytes()
-    from tree_sitter_analyzer.diff_snapshot_capture import ChangedFile
-
-    install_fake_snapshot_materializer(
-        monkeypatch,
-        root,
-        records=[ChangedFile("old.py", "M", True, True, False)],
-        inventory_paths=["old.py"],
-    )
-    observations: list[bytes] = []
-
-    def legacy_dependency_analysis(_project_root):
-        dependency.write_bytes(b"TRANSIENT = True\n")
-        observations.append(dependency.read_bytes())
-        dependency.write_bytes(original)
-        return None
-
-    monkeypatch.setattr(
-        change_impact_analysis, "_load_dependency_graph", legacy_dependency_analysis
-    )
-
-    result = asyncio.run(
-        build_edit_facade(str(root)).execute(
-            {
-                "action": "impact",
-                "mode": "diff",
-                "capture_diff_snapshot": True,
-                "output_format": "json",
-            }
-        )
-    )
-
-    assert observations == []
-    assert dependency.read_bytes() == original
-    assert result["affected_files_unknown"] is True
-
-
-def test_edit_release_snapshot_is_same_process_reachable_and_idempotent(
-    tmp_path: Path, monkeypatch
-) -> None:
-    # PR #1252 review thread 3746878592.
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    root = make_repo(tmp_path)
-    (root / "old.py").write_text("value = 2\n")
-    install_fake_snapshot_materializer(monkeypatch, root)
-    registry = snapshots.DiffSnapshotRegistry()
-    monkeypatch.setattr(snapshots, "REGISTRY", registry)
-    created = registry.create(str(root), "diff", [])
-    args = {
-        "action": "release_snapshot",
-        "diff_snapshot_id": created["diff_snapshot_id"],
-        "route_lease_id": created["route_lease_id"],
-        "output_format": "json",
-    }
-    facade = build_edit_facade(str(root))
-
-    first = asyncio.run(facade.execute(args))
-    second = asyncio.run(facade.execute(args))
-
-    assert (first["released"], second["released"]) == (True, True)
-
-
-def test_edit_release_snapshot_rejects_wrong_ownership_token(
-    tmp_path: Path, monkeypatch
-) -> None:
-    # PR #1252 review thread 3746878592.
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    root = make_repo(tmp_path)
-    install_fake_snapshot_materializer(monkeypatch, root)
-    registry = snapshots.DiffSnapshotRegistry()
-    monkeypatch.setattr(snapshots, "REGISTRY", registry)
-    created = registry.create(str(root), "diff", [])
-
-    result = asyncio.run(
-        build_edit_facade(str(root)).execute(
-            {
-                "action": "release_snapshot",
-                "diff_snapshot_id": created["diff_snapshot_id"],
-                "route_lease_id": "wrong",
-                "output_format": "json",
-            }
-        )
-    )
-
-    assert result["error_code"] == "DIFF_SNAPSHOT_LEASE_MISMATCH"
-
-
-def test_edit_release_snapshot_rejects_alternate_source_arguments() -> None:
-    # PR #1252 review thread 3746878592.
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    with pytest.raises(ValueError, match="DIFF_SNAPSHOT_CONFLICTING_ARGUMENTS"):
-        asyncio.run(
-            build_edit_facade(".").execute(
-                {
-                    "action": "release_snapshot",
-                    "diff_snapshot_id": "ds",
-                    "route_lease_id": "lease",
-                    "file_path": "alternate.py",
-                }
-            )
-        )
-
-
-def test_edit_release_snapshot_requires_both_ownership_ids() -> None:
-    # PR #1252 review thread 3746878592.
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    with pytest.raises(
-        ValueError, match="diff_snapshot_id and route_lease_id are required"
-    ):
-        asyncio.run(
-            build_edit_facade(".").execute(
-                {"action": "release_snapshot", "diff_snapshot_id": "ds"}
-            )
-        )
-
-
-def test_change_impact_annotation_is_non_idempotent_for_optional_capture() -> None:
-    # PR #1252 review thread 3747113064.
-    from tree_sitter_analyzer.mcp.tools.change_impact_tool import ChangeImpactTool
-
-    definition = ChangeImpactTool().get_tool_definition()
-    assert definition["annotations"]["idempotentHint"] is False
-
-
-@pytest.mark.asyncio
-async def test_edit_constraints_projects_exact_frozen_scope_arguments(
-    monkeypatch,
-) -> None:
-    from tree_sitter_analyzer.mcp.tools.constraint_check_tool import ConstraintCheckTool
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    seen: list[dict[str, object]] = []
-
-    async def fake_execute(self, arguments):
-        seen.append(dict(arguments))
-        return {"success": True}
-
-    monkeypatch.setattr(ConstraintCheckTool, "execute", fake_execute)
-    await build_edit_facade(None).execute(
-        {
-            "action": "constraints",
-            "persist": False,
-            "diff_snapshot_id": "ds_contract",
-            "scope_paths": ["src/a.py", "src/b.py"],
-            "output_format": "json",
-        }
-    )
-
-    assert seen == [
-        {
-            "persist": False,
-            "diff_snapshot_id": "ds_contract",
-            "scope_paths": ["src/a.py", "src/b.py"],
-            "output_format": "json",
-        }
-    ]
-
-
-def test_edit_constraints_snapshot_parameters_are_schema_discoverable() -> None:
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    properties = build_edit_facade(None).get_tool_definition()["inputSchema"][
-        "properties"
-    ]
-
-    assert properties["persist"] == {
-        "type": "boolean",
-        "default": True,
-        "description": (
-            "Write evaluated violations through to the cache. Set false for "
-            "RFC-0022 read-only evaluation; no database or file is created."
-        ),
-    }
-    assert properties["scope_paths"] == {
-        "type": "array",
-        "items": {"type": "string"},
-        "description": (
-            "Primitive-issued frozen scope for action=constraints, or impact "
-            "capture scope for action=impact."
-        ),
-    }
-
-
-@pytest.mark.asyncio
-async def test_edit_constraints_rejects_snapshot_without_persist_false() -> None:
-    from tree_sitter_analyzer.mcp.tools.edit_facade import build_edit_facade
-
-    with pytest.raises(ValueError, match="diff_snapshot_id requires persist=false"):
-        await build_edit_facade(None).execute(
-            {
-                "action": "constraints",
-                "diff_snapshot_id": "ds_contract",
-                "scope_paths": ["src/a.py"],
-            }
-        )
