+++ tests/unit/test_ast_diff_tool.py
@@ -2,44 +2,11 @@
 
 from __future__ import annotations
 
-import asyncio
-from pathlib import Path
 from unittest.mock import MagicMock, patch
 
 import pytest
 
-import tree_sitter_analyzer.diff_snapshot_epoch as epoch_verification
-import tree_sitter_analyzer.diff_snapshot_registry as snapshots
-from tests.unit._diff_snapshot_support import POSIX_SNAPSHOT_TEST, make_repo
 from tree_sitter_analyzer.mcp.tools.ast_diff_tool import ASTDiffTool
-from tree_sitter_analyzer.mcp.tools.semantic_classify_tool import SemanticClassifyTool
-
-
-def _create_stable_consumer_snapshot(
-    monkeypatch: pytest.MonkeyPatch, root: Path
-) -> dict[str, object]:
-    """Capture once, then make consumer publication verification deterministic."""
-    monkeypatch.setattr(
-        epoch_verification.FrozenGitEnvironment,
-        "verify_source_epoch",
-        lambda self: None,
-    )
-    created = snapshots.REGISTRY.create(str(root), "diff", [])
-    identity = snapshots.canonical_root(str(root))[1]
-    generation = str(created["source_generation"])
-    state = snapshots.REGISTRY._states[str(created["diff_snapshot_id"])]
-    git_generation = state.snapshot.git_generation
-    monkeypatch.setattr(
-        snapshots,
-        "oracle_generation",
-        lambda project_root, mode="diff", *, deadline=None: (git_generation, identity),
-    )
-    monkeypatch.setattr(
-        snapshots,
-        "shared_source_generation",
-        lambda *_args, **_kwargs: generation,
-    )
-    return created
 
 
 @pytest.fixture
@@ -386,150 +353,6 @@ _NEW_SRC = "def foo():\n    pass\n\ndef bar():\n    pass\n"
 _LANG = "python"
 
 
-class TestASTDiffNodeBudget:
-    """Issue #552 — default response must NOT contain children; opt-in adds them."""
-
-    @pytest.mark.asyncio
-    async def test_default_response_has_no_children_key(self, tool):
-        """Hunk nodes must NOT contain a 'children' key when include_node_bodies is omitted."""
-        result = await tool.execute(
-            {
-                "mode": "diff_strings",
-                "old_source": _OLD_SRC,
-                "new_source": _NEW_SRC,
-                "language": _LANG,
-                "output_format": "json",
-            }
-        )
-        hunks = result.get("hunks", [])
-        assert len(hunks) == 1  # exactly one added hunk
-        hunk = hunks[0]
-        new_node = hunk.get("new")
-        assert new_node is not None, "hunk must have 'new' node"
-        assert "children" not in new_node, (
-            "default response must NOT inline children — issue #552"
-        )
-
-    @pytest.mark.asyncio
-    async def test_default_response_has_child_count(self, tool):
-        """Hunk nodes MUST have child_count (exact pin) when include_node_bodies is omitted."""
-        result = await tool.execute(
-            {
-                "mode": "diff_strings",
-                "old_source": _OLD_SRC,
-                "new_source": _NEW_SRC,
-                "language": _LANG,
-                "output_format": "json",
-            }
-        )
-        hunks = result.get("hunks", [])
-        assert len(hunks) == 1
-        new_node = hunks[0].get("new", {})
-        assert "child_count" in new_node, "child_count must be present in default mode"
-        # Grammar-pinned exact assertion: function_definition has 5 children
-        # (def keyword, name identifier, parameters, colon, body block).
-        # If a grammar bump shifts this, the test MUST go red.
-        assert new_node["child_count"] == 5
-
-    @pytest.mark.asyncio
-    async def test_schema_has_include_node_bodies_param(self, tool):
-        """Schema must expose include_node_bodies boolean param."""
-        schema = tool.get_tool_schema()
-        props = schema.get("properties", {})
-        assert "include_node_bodies" in props, (
-            "include_node_bodies param must be in the tool schema"
-        )
-        assert props["include_node_bodies"].get("type") == "boolean"
-        # Must NOT be in required (runtime-resolved param convention)
-        assert "include_node_bodies" not in schema.get("required", [])
-
-    @pytest.mark.asyncio
-    async def test_default_bytes_smaller_than_include_bodies_bytes(self, tool):
-        """DOCUMENTED RELATIONSHIP: default response < include_node_bodies=True response."""
-        import json
-
-        args_base = {
-            "mode": "diff_strings",
-            "old_source": _OLD_SRC,
-            "new_source": _NEW_SRC,
-            "language": _LANG,
-            "output_format": "json",
-        }
-        default_result = await tool.execute(args_base)
-        bodies_result = await tool.execute({**args_base, "include_node_bodies": True})
-
-        # Strip volatile envelope fields (timing varies run-to-run) so the
-        # byte counts are deterministic and can be pinned EXACTLY (CLAUDE.md
-        # locked rule: no loose </> assertions — a grammar bump SHOULD go red
-        # and force a conscious re-pin).
-        _volatile = {
-            "elapsed_ms",
-            "cache_age_s",
-            "from_cache",
-            "cache_invalidated_reason",
-        }
-
-        def _stable_bytes(d):
-            return len(json.dumps({k: v for k, v in d.items() if k not in _volatile}))
-
-        default_bytes = _stable_bytes(default_result)
-        bodies_bytes = _stable_bytes(bodies_result)
-
-        # Cost invariant (CLAUDE.md rule 11) pinned exactly: 746 < 1877 — the
-        # default is dramatically smaller than the full-body opt-in.
-        # Re-pinned after Codex P2 made agent_summary a dict (next_step+verdict).
-        assert default_bytes == 746
-        assert bodies_bytes == 1877
-
-    @pytest.mark.asyncio
-    async def test_include_node_bodies_true_has_children(self, tool):
-        """When include_node_bodies=True, hunk nodes MUST contain 'children' key."""
-        result = await tool.execute(
-            {
-                "mode": "diff_strings",
-                "old_source": _OLD_SRC,
-                "new_source": _NEW_SRC,
-                "language": _LANG,
-                "output_format": "json",
-                "include_node_bodies": True,
-            }
-        )
-        hunks = result.get("hunks", [])
-        assert len(hunks) == 1
-        new_node = hunks[0].get("new", {})
-        assert "children" in new_node, (
-            "include_node_bodies=True must inline children in hunk nodes"
-        )
-        # Exact pin: function_definition has 5 children (grammar-pinned)
-        assert len(new_node["children"]) == 5
-
-    @pytest.mark.asyncio
-    async def test_over_budget_sets_children_truncated(self, tool):
-        """When include_node_bodies=True and response exceeds budget, set children_truncated."""
-        # Patch the budget constant to 1 byte to guarantee truncation
-        with patch(
-            "tree_sitter_analyzer.mcp.tools.ast_diff_tool.NODE_BODIES_BUDGET", 1
-        ):
-            result = await tool.execute(
-                {
-                    "mode": "diff_strings",
-                    "old_source": _OLD_SRC,
-                    "new_source": _NEW_SRC,
-                    "language": _LANG,
-                    "output_format": "json",
-                    "include_node_bodies": True,
-                }
-            )
-        # Must set the transparency flag when budget is exceeded
-        assert result.get("children_truncated") is True, (
-            "children_truncated must be True when budget is exceeded"
-        )
-        assert "bytes_omitted" in result, "bytes_omitted must be present when truncated"
-        # Exact pin (deterministic for this fixture at budget=1): the omitted
-        # bytes equal full-body minus compact = 1627 - 496 = 1131.
-        assert result["bytes_omitted"] == 1131
-
-
 class TestAstDiffAgentSummaryEnvelope:
     """#744: ast_diff must include agent_summary and summary_line in response."""
 
@@ -594,388 +417,3 @@ class TestAstDiffAgentSummaryEnvelope:
         assert result["verdict"] == "ERROR"
         assert result["agent_summary"]["verdict"] == "ERROR"
         assert result["agent_summary"]["summary_line"] == "Both sources failed to parse"
-
-
-@pytest.mark.asyncio
-async def test_snapshot_requires_file_path(tool) -> None:
-    with pytest.raises(ValueError, match="DIFF_SNAPSHOT_FILE_REQUIRED"):
-        await tool.execute({"diff_snapshot_id": "ds"})
-
-
-@pytest.mark.asyncio
-async def test_snapshot_translates_registry_error(tool, monkeypatch) -> None:
-    from tree_sitter_analyzer import diff_snapshot_registry as registry
-
-    monkeypatch.setattr(
-        registry.REGISTRY, "acquire", lambda *a: (None, "DIFF_SNAPSHOT_EXPIRED")
-    )
-    result = await tool.execute(
-        {"diff_snapshot_id": "ds", "file_path": "x.py", "output_format": "json"}
-    )
-    assert result["error_code"] == "DIFF_SNAPSHOT_EXPIRED"
-
-
-@pytest.mark.asyncio
-async def test_snapshot_reports_missing_frozen_file(tool, monkeypatch) -> None:
-    from types import SimpleNamespace
-
-    from tree_sitter_analyzer import diff_snapshot_registry as registry
-
-    consumer = SimpleNamespace(
-        snapshot=SimpleNamespace(file=lambda path: None), release=lambda: None
-    )
-    monkeypatch.setattr(registry.REGISTRY, "acquire", lambda *a: (consumer, None))
-    result = await tool.execute(
-        {"diff_snapshot_id": "ds", "file_path": "x.py", "output_format": "json"}
-    )
-    assert result["error_code"] == "DIFF_SNAPSHOT_FILE_NOT_FOUND"
-
-
-@pytest.mark.asyncio
-async def test_snapshot_rejects_non_utf8_frozen_bytes(tool, monkeypatch) -> None:
-    from types import SimpleNamespace
-
-    from tree_sitter_analyzer import diff_snapshot_registry as registry
-
-    frozen = SimpleNamespace(
-        record=SimpleNamespace(path="x.py", binary=False),
-        old_bytes=b"\xff",
-        new_bytes=b"",
-    )
-    consumer = SimpleNamespace(
-        snapshot=SimpleNamespace(file=lambda path: frozen), release=lambda: None
-    )
-    monkeypatch.setattr(registry.REGISTRY, "acquire", lambda *a: (consumer, None))
-    result = await tool.execute(
-        {"diff_snapshot_id": "ds", "file_path": "x.py", "output_format": "json"}
-    )
-    assert result["error_code"] == "DIFF_SNAPSHOT_UNSUPPORTED_CONTENT"
-
-
-@pytest.mark.parametrize(
-    ("tool_type", "field", "expected"),
-    [(ASTDiffTool, "hunks", 2), (SemanticClassifyTool, "change_count", 2)],
-)
-@POSIX_SNAPSHOT_TEST
-def test_snapshot_consumer_uses_frozen_utf8_bytes(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    tool_type,
-    field: str,
-    expected: int,
-) -> None:
-    root = make_repo(tmp_path)
-    (root / "old.py").write_text("value = 2\n")
-    result = _create_stable_consumer_snapshot(monkeypatch, root)
-    request = {
-        "diff_snapshot_id": result["diff_snapshot_id"],
-        "file_path": "old.py",
-        "output_format": "json",
-    }
-    response = asyncio.run(tool_type(str(root)).execute(request))
-    assert (
-        len(response[field]) if isinstance(response[field], list) else response[field]
-    ) == expected
-    assert (
-        snapshots.REGISTRY.close_lease(
-            str(result["diff_snapshot_id"]), str(result["route_lease_id"])
-        )
-        is True
-    )
-
-
-@pytest.mark.parametrize("tool_type", [ASTDiffTool, SemanticClassifyTool])
-@pytest.mark.parametrize("output_format", ["json", "toon"])
-@POSIX_SNAPSHOT_TEST
-def test_snapshot_consumer_echoes_exact_frozen_identity(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    tool_type,
-    output_format: str,
-) -> None:
-    # PR #1252 review thread 3748730795: consumers must not infer identity.
-    root = make_repo(tmp_path)
-    (root / "old.py").write_text("value = 2\n")
-    created = _create_stable_consumer_snapshot(monkeypatch, root)
-    response = asyncio.run(
-        tool_type(str(root)).execute(
-            {
-                "diff_snapshot_id": created["diff_snapshot_id"],
-                "file_path": "old.py",
-                "output_format": output_format,
-            }
-        )
-    )
-    assert response["diff_snapshot_id"] == created["diff_snapshot_id"]
-    assert response["source_generation"] == created["source_generation"]
-    assert (
-        snapshots.REGISTRY.close_lease(
-            str(created["diff_snapshot_id"]), str(created["route_lease_id"])
-        )
-        is True
-    )
-
-
-@pytest.mark.parametrize("tool_type", [ASTDiffTool, SemanticClassifyTool])
-@POSIX_SNAPSHOT_TEST
-def test_snapshot_consumer_rejects_binary_content(
-    tmp_path: Path, monkeypatch: pytest.MonkeyPatch, tool_type
-) -> None:
-    root = make_repo(tmp_path)
-    (root / "blob.py").write_bytes(b"a\0b")
-    result = _create_stable_consumer_snapshot(monkeypatch, root)
-    request = {
-        "diff_snapshot_id": result["diff_snapshot_id"],
-        "file_path": "blob.py",
-        "output_format": "json",
-    }
-    response = asyncio.run(tool_type(str(root)).execute(request))
-    assert response["error_code"] == "DIFF_SNAPSHOT_UNSUPPORTED_CONTENT"
-    assert (
-        snapshots.REGISTRY.close_lease(
-            str(result["diff_snapshot_id"]), str(result["route_lease_id"])
-        )
-        is True
-    )
-
-
-@pytest.mark.parametrize(
-    ("tool_type", "module_name"),
-    [
-        (ASTDiffTool, "tree_sitter_analyzer.mcp.tools.ast_diff_tool"),
-        (
-            SemanticClassifyTool,
-            "tree_sitter_analyzer.mcp.tools.semantic_classify_tool",
-        ),
-    ],
-)
-@pytest.mark.parametrize(
-    ("validation_error", "request_format", "expected_error"),
-    [
-        (
-            "DIFF_SNAPSHOT_SOURCE_CHANGED",
-            "toon",
-            "DIFF_SNAPSHOT_SOURCE_CHANGED",
-        ),
-        ("DIFF_SNAPSHOT_EXPIRED", None, "DIFF_SNAPSHOT_EXPIRED"),
-        (
-            "DIFF_SNAPSHOT_FUTURE_ERROR",
-            None,
-            "DIFF_SNAPSHOT_VALIDATION_ERROR",
-        ),
-    ],
-)
-@pytest.mark.asyncio
-async def test_strict_snapshot_final_publish_errors_preserve_toon(
-    tool_type,
-    module_name: str,
-    validation_error: str,
-    request_format: str | None,
-    expected_error: str,
-    monkeypatch,
-) -> None:
-    from importlib import import_module
-    from types import SimpleNamespace
-
-    from tree_sitter_analyzer import diff_snapshot_registry as registry
-
-    events: list[tuple[str, bool]] = []
-    released = False
-    frozen = SimpleNamespace(
-        record=SimpleNamespace(path="x.py", binary=False),
-        old_bytes=b"value = 1\n",
-        new_bytes=b"value = 2\n",
-    )
-
-    def release() -> None:
-        nonlocal released
-        released = True
-        events.append(("release", released))
-
-    consumer = SimpleNamespace(
-        snapshot=SimpleNamespace(file=lambda path: frozen), release=release
-    )
-    monkeypatch.setattr(registry.REGISTRY, "acquire", lambda *a: (consumer, None))
-
-    def validate_publish(pinned) -> str:
-        events.append(("validate", released))
-        return validation_error
-
-    monkeypatch.setattr(registry.REGISTRY, "validate_publish", validate_publish)
-    tool_module = import_module(module_name)
-    original_formatter = tool_module.apply_toon_format_to_response
-
-    def formatting(response, output_format):
-        events.append(("format", released))
-        return original_formatter(response, output_format)
-
-    monkeypatch.setattr(tool_module, "apply_toon_format_to_response", formatting)
-    request = {"diff_snapshot_id": "ds", "file_path": "x.py"}
-    if request_format is not None:
-        request["output_format"] = request_format
-    response = await tool_type(".").execute(request)
-    # PR #1252 review thread 3750964908: final validation must not bypass TOON.
-    assert response["error_code"] == expected_error
-    assert response["format"] == "toon"
-    assert isinstance(response["toon_content"], str)
-    assert events == [("format", False)] * 22 + [
-        ("validate", False),
-        ("release", True),
-    ]
-
-
-def test_snapshot_consumers_reject_symlink_kind() -> None:
-    # PR #1252 review thread 3746878582.
-    from types import SimpleNamespace
-
-    from tree_sitter_analyzer.mcp.tools.ast_diff_tool import ASTDiffTool
-
-    frozen = SimpleNamespace(
-        record=SimpleNamespace(
-            path="module.py", binary=False, old_kind="symlink", new_kind="symlink"
-        ),
-        old_bytes=b"target.py",
-        new_bytes=b"other.py",
-    )
-    consumer = SimpleNamespace(
-        snapshot=SimpleNamespace(file=lambda path: frozen), release=lambda: None
-    )
-    with patch(
-        "tree_sitter_analyzer.diff_snapshot_registry.REGISTRY.acquire",
-        return_value=(consumer, None),
-    ):
-        result = asyncio.run(
-            ASTDiffTool(".").execute(
-                {
-                    "diff_snapshot_id": "ds",
-                    "file_path": "module.py",
-                    "output_format": "json",
-                }
-            )
-        )
-    assert result["error_code"] == "DIFF_SNAPSHOT_UNSUPPORTED_CONTENT"
-
-
-def test_snapshot_ast_options_are_not_source_conflicts() -> None:
-    # PR #1252 review thread 3746878597.
-    from tree_sitter_analyzer.mcp.tools.ast_diff_tool import ASTDiffTool
-
-    assert (
-        ASTDiffTool(".").validate_arguments(
-            {
-                "diff_snapshot_id": "ds",
-                "file_path": "module.py",
-                "include_node_bodies": True,
-                "output_format": "json",
-            }
-        )
-        is True
-    )
-
-
-def test_ast_diff_execute_rejects_unreachable_unknown_mode() -> None:
-    tool = ASTDiffTool(".")
-    with (
-        patch.object(tool, "validate_arguments", return_value=True),
-        patch.object(tool, "_resolve_mode", return_value="unknown"),
-        pytest.raises(ValueError, match="Unknown mode: unknown"),
-    ):
-        asyncio.run(tool.execute({"output_format": "json"}))
-
-
-@pytest.mark.parametrize("tool_type", [ASTDiffTool, SemanticClassifyTool])
-@pytest.mark.parametrize("unavailable_side", ["old_available", "new_available"])
-@pytest.mark.asyncio
-async def test_snapshot_consumer_rejects_each_unavailable_side(
-    tool_type, unavailable_side: str, monkeypatch
-) -> None:
-    # PR #1252 review thread 3748259951.
-    from types import SimpleNamespace
-
-    from tree_sitter_analyzer import diff_snapshot_registry as registry
-
-    availability = {"old_available": True, "new_available": True}
-    availability[unavailable_side] = False
-    frozen = SimpleNamespace(
-        record=SimpleNamespace(path="x.py", binary=False, **availability),
-        old_bytes=b"value = 1\n",
-        new_bytes=b"value = 2\n",
-    )
-    consumer = SimpleNamespace(
-        snapshot=SimpleNamespace(file=lambda path: frozen), release=lambda: None
-    )
-    monkeypatch.setattr(registry.REGISTRY, "acquire", lambda *a: (consumer, None))
-
-    result = await tool_type(".").execute(
-        {"diff_snapshot_id": "ds", "file_path": "x.py", "output_format": "json"}
-    )
-
-    assert result["error_code"] == "DIFF_SNAPSHOT_UNSUPPORTED_CONTENT"
-
-
-@pytest.mark.parametrize("tool_type", [ASTDiffTool, SemanticClassifyTool])
-@pytest.mark.parametrize("status", ["R", "C"])
-@pytest.mark.asyncio
-async def test_snapshot_consumer_rejects_rename_and_copy_status(
-    tool_type, status: str, monkeypatch
-) -> None:
-    # PR #1252 review thread 3748575979.
-    from types import SimpleNamespace
-
-    frozen = SimpleNamespace(
-        record=SimpleNamespace(path="x.py", status=status, binary=False),
-        old_bytes=b"value = 1\n",
-        new_bytes=b"value = 2\n",
-    )
-    consumer = SimpleNamespace(
-        snapshot=SimpleNamespace(file=lambda path: frozen), release=lambda: None
-    )
-    monkeypatch.setattr(snapshots.REGISTRY, "acquire", lambda *a: (consumer, None))
-    result = await tool_type(".").execute(
-        {"diff_snapshot_id": "ds", "file_path": "x.py", "output_format": "json"}
-    )
-    assert result["error_code"] == "DIFF_SNAPSHOT_UNSUPPORTED_CONTENT"
-
-
-@pytest.mark.parametrize("tool_type", [ASTDiffTool, SemanticClassifyTool])
-def test_snapshot_consumer_rejects_caller_language_override(tool_type) -> None:
-    # PR #1252 review thread 4873: strict language is bound to captured path.
-    with pytest.raises(ValueError, match="DIFF_SNAPSHOT_CONFLICTING_ARGUMENTS"):
-        tool_type(".").validate_arguments(
-            {
-                "diff_snapshot_id": "ds",
-                "file_path": "module.py",
-                "language": "javascript",
-            }
-        )
-
-
-@pytest.mark.parametrize("tool_type", [ASTDiffTool, SemanticClassifyTool])
-@pytest.mark.asyncio
-async def test_snapshot_consumer_rejects_unknown_captured_extension(
-    tool_type, monkeypatch
-) -> None:
-    # PR #1252 review thread 4873: unknown captured extension is stable unsupported.
-    from types import SimpleNamespace
-
-    from tree_sitter_analyzer import diff_snapshot_registry as registry
-
-    frozen = SimpleNamespace(
-        record=SimpleNamespace(path="module.unknown", binary=False),
-        old_bytes=b"old",
-        new_bytes=b"new",
-    )
-    consumer = SimpleNamespace(
-        snapshot=SimpleNamespace(file=lambda path: frozen), release=lambda: None
-    )
-    monkeypatch.setattr(registry.REGISTRY, "acquire", lambda *a: (consumer, None))
-
-    result = await tool_type(".").execute(
-        {
-            "diff_snapshot_id": "ds",
-            "file_path": "module.unknown",
-            "output_format": "json",
-        }
-    )
-
-    assert result["error_code"] == "DIFF_SNAPSHOT_UNSUPPORTED_LANGUAGE"
